refactored telemetry
