+++ digitalpy/core/telemetry/metrics.py
@@ -1,0 +1,4 @@
+from digitalpy.core.zmanager.controller import Controller
+
+class Metrics(Controller):
+    pass
