--- digitalpy/telemetry/metrics.py
@@ -1,4 +1,0 @@
-from digitalpy.routing.controller import Controller
-
-class Metrics(Controller):
-    pass
